Calculate the accuracy of the model in frugally-deep
Run all predictions in the first training file of the cifar10 dataset using
frugally and calculate the accuracy of said predictions. With the current
model, accuracy is 10%, meaning its just as good as random classification. This
can likely be fixed by training a "real" model i.e. with multiple epochs and on
the same train/test set separation.